Rename `lib/logger` to `lib/factory` for clarity
