+++ lib/factory.js
